develop<db>: entity QuizResultAnswer
