+++ src/MasterPeace/Bundle/QuizBundle/Repository/QuizResultAnswerRepository.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace MasterPeace\Bundle\QuizBundle\Repository;
+
+/**
+ * QuizResultAnswerRepository
+ *
+ * This class was generated by the Doctrine ORM. Add your own custom
+ * repository methods below.
+ */
+class QuizResultAnswerRepository extends \Doctrine\ORM\EntityRepository
+{
+}
